+++ src/ready_late.lua
@@ -6,4 +6,21 @@ modutil.mod.Path.Context.Env("OpenRunClearScreen", function ()
         end
         return base()
     end)
+    modutil.mod.Path.Wrap("SetAlpha", function (base, args)
+        local screen = game.ActiveScreens.RunClear
+        if screen and screen[_PLUGIN.guid .. "SkipRecordRunCleared"] and args.Id == screen.Components.ClearTimeRecord.Id then
+            return
+        end
+        return base(args)
+    end)
+end)
+
+modutil.mod.Path.Context.Env("KillHero", function ()
+    modutil.mod.Path.Wrap("LoadMap", function (base, args)
+        args = args or {}
+        if game.CurrentRun[_PLUGIN.guid .. "Retry"] and game.CurrentRun[_PLUGIN.guid .. "SavedStartOverArgs"] then
+            return game.StartOver(game.CurrentRun[_PLUGIN.guid .. "SavedStartOverArgs"])
+        end
+        return base(args)
+    end)
 end)